# Remove double lookups on SizeLimitedCache<K, V>, improve performance