--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/MS/internal/Shaping/GlyphingCache.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/MS/internal/Shaping/GlyphingCache.cs
@@ -30,7 +30,7 @@ namespace MS.Internal.Shaping
     /// </summary>
     internal class GlyphingCache
     {
-        private SizeLimitedCache<Typeface, TypefaceMap>  _sizeLimitedCache;
+        private readonly SizeLimitedCache<Typeface, TypefaceMap> _sizeLimitedCache;
     
         internal GlyphingCache(int capacity)
         {