--- a/src/Microsoft.DotNet.Wpf/src/Shared/MS/Internal/SizeLimitedCache.cs
+++ b/src/Microsoft.DotNet.Wpf/src/Shared/MS/Internal/SizeLimitedCache.cs
@@ -85,20 +85,13 @@ namespace MS.Internal
         /// </param>
         public void Add(K key, V resource, bool isPermanent)
         {
-
-            if ( (object)key == null)
-            {
-                throw new ArgumentNullException(nameof(key));
-            }
-            if ( (object)resource == null)
-            {
-                throw new ArgumentNullException(nameof(resource));
-            }
+            ArgumentNullException.ThrowIfNull(key, nameof(key));
+            ArgumentNullException.ThrowIfNull(resource, nameof(resource));
 
             // note: [] throws, thus we should check if its in the dictionary first.
-            if (!_nodeLookup.ContainsKey(key))
+            if (!_nodeLookup.TryGetValue(key, out Node node))
             {
-                Node node = new Node(key, resource, isPermanent);
+                node = new Node(key, resource, isPermanent);
                 if (!isPermanent)
                 {
                     if (IsFull())
@@ -111,11 +104,11 @@ namespace MS.Internal
                 {
                     _permanentCount++;
                 }
+
                 _nodeLookup[key] = node;
             }
             else
             {
-                Node node = _nodeLookup[key];
                 if (!node.IsPermanent)
                 {
                     RemoveFromList(node);
@@ -151,27 +144,17 @@ namespace MS.Internal
         /// </param>
         public void Remove(K key)
         {
-            if ( (object)key == null)
-            {
-                throw new ArgumentNullException(nameof(key));
-            }
+            ArgumentNullException.ThrowIfNull(key, nameof(key));
 
-            // note: [] throws, thus we should check if its in the dictionary first.
-            if (!_nodeLookup.ContainsKey(key))
-            {
+            if (!_nodeLookup.TryGetValue(key, out Node node))
                 return;
-            }
-            Node node = _nodeLookup[key];
 
             _nodeLookup.Remove(key);
+
             if (!node.IsPermanent)
-            {
                 RemoveFromList(node);
-            }
             else
-            {
                 _permanentCount--;
-            }
         }
 
         /// <summary>