Fixes #4828 

## Description

While checking out `SizeLimitedCache<K, V>` yesterday, I've noticed the double lookups on the dictionary and today I've found there's already a corresponding issue for it, hence we should fix it for optimal performance as this is used with glyphs/fonts.

- Use throw helpers over creating new instance of `ArgumentNullException` directly.
- `Remove` had even three lookups, so now it just has one.
- Fields that could be `readonly` have been adjusted.
- Sealed the class, proper naming conventions.
- `Node` uses auto properties now.

### 10 capacity, 15 unique additions first, 20 lookups

| Method   | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|--------- |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original |  988.4 ns |    9.39 ns |     7.84 ns | 0.1183 |       2,045 B |        1992 B |
| PR_EDIT  |  809.3 ns |    9.18 ns |     8.14 ns | 0.1183 |       1,820 B |        1992 B |

## Customer Impact

Improved perfomance.

## Regression

No.

## Testing

Local build, tests with the class.

## Risk

Low, just basically compiler warnings.

 ###### Microsoft Reviewers: [Open in CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/dotnet/wpf/pull/9949)